Episode 24 - Authorization With Policies

+++ app/Http/Controllers/ThreadController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Channel;
 use App\Filters\ThreadFilters;
 use App\Http\Requests\StoreThreadRequest;
+use App\Reply;
 use App\Thread;
 use Illuminate\Support\Facades\Log;
 
@@ -80,17 +81,17 @@ class ThreadController extends Controller
      *
      * @param \App\Thread $thread
      * @return \Illuminate\Http\RedirectResponse
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy(Thread $thread)
     {
+        $this->authorize('forceDelete', $thread);
+
         try {
+            $thread->replies()->delete();
             $thread->delete();
         } catch (\Exception $e) {
             Log::error('Failed to delete the thread.');
-        }
-
-        if (request()->wantsJson()) {
-            return response([], 204);
         }
 
         return redirect(route('threads.index'));
